Initial pass to fix spellings/grammar etc.
For the most part, simply touching up spellings or lessening repetition of words to improve the flow and last but not least; grammer (sic!). 

Tried to remove gender where possible to either account for the PC's gender, or the fact that there isn't complete certainty over it.

Took a liberty on the stickup scene and tried to give a slightly more old school gangster feel.